+++ packages/environment/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @enzymefinance/environment
 
+## 6.3.79
+
+### Patch Changes
+
+- [#813](https://github.com/enzymefinance/sdk/pull/813) [`140efac`](https://github.com/enzymefinance/sdk/commit/140efacfd87b8388dab517300e215f6474362043) Thanks [@saurabhfegade](https://github.com/saurabhfegade)! - Deregister xUSD on arbitrum
+
+## 6.3.78
+
+### Patch Changes
+
+- [#817](https://github.com/enzymefinance/sdk/pull/817) [`ac01a00`](https://github.com/enzymefinance/sdk/commit/ac01a00e252090f3788231c67c780e48841af960) Thanks [@gabrocheleau](https://github.com/gabrocheleau)! - Add bebop blend to exchanges
+
+## 6.3.77
+
+### Patch Changes
+
+- [#815](https://github.com/enzymefinance/sdk/pull/815) [`eed2815`](https://github.com/enzymefinance/sdk/commit/eed2815c931484c1b985c6f209166d92702fd2ca) Thanks [@gabrocheleau](https://github.com/gabrocheleau)! - Add BebopBlendAdapter sdk methods and testnet deployment address
+
+- Updated dependencies [[`eed2815`](https://github.com/enzymefinance/sdk/commit/eed2815c931484c1b985c6f209166d92702fd2ca)]:
+  - @enzymefinance/sdk@1.2.27
+
+## 6.3.76
+
+### Patch Changes
+
+- [#809](https://github.com/enzymefinance/sdk/pull/809) [`75eee15`](https://github.com/enzymefinance/sdk/commit/75eee1557d4521d7ffa00f0a6809dc1485f9d03c) Thanks [@saurabhfegade](https://github.com/saurabhfegade)! - Add Paxos Gold and BUIDL on Arbitrum
+
+## 6.3.75
+
+### Patch Changes
+
+- [#811](https://github.com/enzymefinance/sdk/pull/811) [`d0313c3`](https://github.com/enzymefinance/sdk/commit/d0313c3cab01e9a3e10a543880427bc86b147b8c) Thanks [@gabrocheleau](https://github.com/gabrocheleau)! - Add USDC to named tokens
+
 ## 6.3.74
 
 ### Patch Changes